refactor(components): 表示用 Avatar 基底を共有パッケージの display へ昇格
app 層に埋もれていた app 非依存の表示用 Avatar を shared へ昇格し、
新 taxonomy の display カテゴリに配置する。

- 基底 Avatar/AvatarImage/AvatarFallback + avatarVariants を
  packages/components/src/display/avatar.tsx へ移設し barrel から export
- @radix-ui/react-avatar を packages/components の依存に追加
- app 固有の AvatarUpload（Supabase Storage アップロード前提）は
  components/ui/avatar-upload.tsx に残し、基底は @dayopt/components から取得
- 基底 Avatar の consumer 4 箇所を @dayopt/components へ切替
- settings routing test の mock を @dayopt/components の partial mock へ更新

typecheck / lint / lint:boundaries / 全 1848 test / build pass。
