+++ apps/product/src/app/[locale]/(app)/settings/__tests__/routing.test.tsx
@@ -72,7 +72,10 @@ vi.mock('@/lib/hooks/useLogout', () => ({
   useLogout: () => ({ logout: mockLogout, isLoggingOut: false }),
 }));
 
-vi.mock('@/components/ui/avatar', () => ({
+// 基底 Avatar は @dayopt/components に昇格済み。表示系のみ stub し、その他の
+// 共有 component（Card / Badge / ScrollArea / Skeleton 等）は実体を使う。
+vi.mock('@dayopt/components', async (importOriginal) => ({
+  ...(await importOriginal<typeof import('@dayopt/components')>()),
   Avatar: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
   AvatarImage: () => null,
   AvatarFallback: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
