+++ apps/product/src/components/ui/avatar-upload.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import * as AvatarPrimitive from '@radix-ui/react-avatar';
-import { cva, type VariantProps } from 'class-variance-authority';
 import { Camera, Loader2, Trash2, Upload, User } from 'lucide-react';
 import { useTranslations } from 'next-intl';
 import Image from 'next/image';
@@ -10,79 +8,7 @@ import { useCallback, useMemo, useState } from 'react';
 import { type FileRejection, useDropzone } from 'react-dropzone';
 
 import { cn } from '@/lib/utils';
-import { Button } from '@dayopt/components';
-
-/**
- * アバターサイズ定義
- *
- * ## サイズ設計（8pxグリッド準拠）
- *
- * | size    | サイズ | 用途                                         |
- * |---------|--------|----------------------------------------------|
- * | xs      | 20px   | インライン、コンパクトリスト                 |
- * | sm      | 32px   | コメント、通知                               |
- * | default | 40px   | 標準的なUI                                   |
- * | lg      | 48px   | プロフィールカード                           |
- * | xl      | 64px   | プロフィールページ、ヒーロー                 |
- * | 2xl     | 96px   | プロフィール編集                             |
- * | 3xl     | 120px  | アバター変更ダイアログ                       |
- */
-const avatarVariants = cva('relative flex shrink-0 overflow-hidden rounded-full', {
-  variants: {
-    size: {
-      xs: 'size-5',
-      sm: 'size-8',
-      default: 'size-10',
-      lg: 'size-12',
-      xl: 'size-16',
-      '2xl': 'size-24',
-      '3xl': 'size-30',
-    },
-  },
-  defaultVariants: {
-    size: 'default',
-  },
-});
-
-interface AvatarProps
-  extends React.ComponentProps<typeof AvatarPrimitive.Root>, VariantProps<typeof avatarVariants> {}
-
-function Avatar({ className, size, ...props }: AvatarProps) {
-  return (
-    <AvatarPrimitive.Root
-      data-slot="avatar"
-      className={cn(avatarVariants({ size }), className)}
-      {...props}
-    />
-  );
-}
-
-function AvatarImage({ className, ...props }: React.ComponentProps<typeof AvatarPrimitive.Image>) {
-  return (
-    <AvatarPrimitive.Image
-      data-slot="avatar-image"
-      className={cn('aspect-square size-full', className)}
-      {...props}
-    />
-  );
-}
-
-function AvatarFallback({
-  className,
-  ...props
-}: React.ComponentProps<typeof AvatarPrimitive.Fallback>) {
-  return (
-    <AvatarPrimitive.Fallback
-      data-slot="avatar-fallback"
-      className={cn('bg-muted flex size-full items-center justify-center rounded-full', className)}
-      {...props}
-    />
-  );
-}
-
-// =============================================================================
-// AvatarUpload - アバター画像アップロード専用コンポーネント
-// =============================================================================
+import { avatarVariants, Button } from '@dayopt/components';
 
 interface AvatarUploadProps {
   /** 現在のアバターURL */
@@ -110,6 +36,9 @@ interface AvatarUploadProps {
  * - 単一画像のみ
  * - 円形プレビュー
  * - ドラッグ&ドロップ対応
+ *
+ * 基底 Avatar（表示用）は @dayopt/components に昇格済み。本コンポーネントは
+ * Supabase Storage アップロード前提の app 固有合成物のため product 側に残す。
  *
  * @example
  * ```tsx
@@ -334,4 +263,4 @@ function formatBytes(bytes: number): string {
   return `${parseFloat((bytes / Math.pow(k, i)).toFixed(0))} ${sizes[i]}`;
 }
 
-export { Avatar, AvatarFallback, AvatarImage, AvatarUpload };
+export { AvatarUpload };
