+++ packages/components/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@radix-ui/react-alert-dialog": "^1.1.16",
+    "@radix-ui/react-avatar": "^1.2.0",
     "@radix-ui/react-checkbox": "^1.3.4",
     "@radix-ui/react-collapsible": "^1.1.13",
     "@radix-ui/react-dialog": "^1.1.16",
